+++ .gitignore
@@ -3,4 +3,4 @@ __pycache__/
 .vscode/
 Results/
 Data
-03-Residual-PA-Unet/wandb/
+wandb/
